+++ src/main/kotlin/org/deniskusakin/aem/groovyconsoleplugin/dsl/AemScriptExtensionClassFinder.kt
@@ -10,41 +10,46 @@ import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.psi.NonClasspathClassFinder
 import com.intellij.psi.PsiClass
 import com.intellij.psi.search.GlobalSearchScope
+import com.intellij.psi.search.NonClasspathDirectoriesScope
 
 /**
  * User: Andrey Bardashevsky
  * Date/Time: 01.08.2022 19:24
  *
- * Hack to provide Node/Page Builders in groovy editor classpath, so groovy plugin able to parse them and highlight builders
  */
 class AemScriptExtensionClassFinder(project: Project) : NonClasspathClassFinder(project) {
 
-    private val everythingScope =  GlobalSearchScope.everythingScope(project)
-
-    override fun calcClassRoots(): List<VirtualFile> {
-        val jarForClass = PathManager.getJarForClass(NodeBuilder::class.java)
-
-        if (jarForClass != null) {
-            val virtualFile = VfsUtil.findFileByIoFile(jarForClass.toFile(), true)
-            if (virtualFile != null) {
-                val classRoot = JarFileSystem.getInstance().getRootByLocal(virtualFile)
-
-                if (classRoot != null) {
-                    return listOf(classRoot)
-                }
-            }
-        }
-
-        return emptyList()
-    }
+    override fun calcClassRoots(): List<VirtualFile> = roots
 
     override fun findClass(qualifiedName: String, scope: GlobalSearchScope): PsiClass? {
         val packageName = StringUtil.getPackageName(qualifiedName)
-        
+
         return if (NodeBuilder::class.java.`package`.name == packageName) {
-            super.findClass(qualifiedName, everythingScope)
+            super.findClass(qualifiedName, scope.uniteWith(searchScope))
         } else {
-            return null
+            null
+        }
+    }
+
+    companion object {
+        private val roots = buildClassesRoots()
+
+        val searchScope = NonClasspathDirectoriesScope.compose(roots)
+
+        private fun buildClassesRoots(): List<VirtualFile> {
+            val jarForClass = PathManager.getJarForClass(NodeBuilder::class.java)
+
+            if (jarForClass != null) {
+                val virtualFile = VfsUtil.findFileByIoFile(jarForClass.toFile(), true)
+                if (virtualFile != null) {
+                    val classRoot = JarFileSystem.getInstance().getRootByLocal(virtualFile)
+
+                    if (classRoot != null) {
+                        return listOf(classRoot)
+                    }
+                }
+            }
+            return emptyList()
         }
     }
 }